# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Companies",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/stations": {
            "get": {
                "tags": [
                    "Stations"
                ],
                "summary": "Get Stations",
                "operationId": "4720e59d102825141a63cf3349253ea0",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "example name"
                                                    },
                                                    "latitude": {
                                                        "type": "number",
                                                        "example": "54.67"
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "example": "43.43"
                                                    },
                                                    "company_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "example address"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Stations"
                ],
                "summary": "Create Station",
                "operationId": "f64cd740423f20d1b8fb367b4e6d0ec2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "example name"
                                            },
                                            "latitude": {
                                                "type": "number",
                                                "example": "54.67"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "example": "43.43"
                                            },
                                            "company_id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "example address"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "latitude": "54.34",
                                    "longitude": "76.98",
                                    "company_id": "1",
                                    "address": "example address"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "title"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "example": "name"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "example": "longitude"
                                        },
                                        "company_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "address"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stations/{id}": {
            "get": {
                "tags": [
                    "Stations"
                ],
                "summary": "Get Station",
                "operationId": "368911e17e985f1cf255154b55ec5c20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "example": "34.54"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "example": "14.65"
                                        },
                                        "company_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "address"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Stations"
                ],
                "summary": "Update Station",
                "operationId": "bd5d2d62515bf9fb50645576008ccc94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "example name"
                                            },
                                            "latitude": {
                                                "type": "number",
                                                "example": "54.67"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "example": "43.43"
                                            },
                                            "company_id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "example address"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "example name",
                                    "latitude": "54.34",
                                    "longitude": "76.98",
                                    "company_id": "1",
                                    "address": "example address"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "latitude": {
                                            "type": "number",
                                            "example": "34.54"
                                        },
                                        "longitude": {
                                            "type": "number",
                                            "example": "14.65"
                                        },
                                        "company_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "address"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stations"
                ],
                "summary": "Delete Station",
                "operationId": "ae4854e3e831ada274aad1d97f5347ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete station success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}